+++ packages/version/index.js
@@ -1,5 +1,5 @@
 module.exports = {
   "decorator": "3.14.4",
   "core": "3.14.4",
-  "timestamp": 1706431995545
+  "timestamp": 1706888765181
 };