+++ scripts/rollback/rollback-3.14.10.sh
@@ -1,0 +1,48 @@
+#!/bin/bash
+# timestamp: 1706888716748
+
+npm dist-tag add @midwayjs/axios@3.14.9 latest
+npm dist-tag add @midwayjs/bull-board@3.14.9 latest
+npm dist-tag add @midwayjs/code-dye@3.14.7 latest
+npm dist-tag add @midwayjs/consul@3.14.7 latest
+npm dist-tag add @midwayjs/cos@3.14.9 latest
+npm dist-tag add @midwayjs/cross-domain@3.14.7 latest
+npm dist-tag add @midwayjs/express-session@3.14.7 latest
+npm dist-tag add @midwayjs/http-proxy@3.14.9 latest
+npm dist-tag add @midwayjs/i18n@3.14.7 latest
+npm dist-tag add @midwayjs/info@3.14.7 latest
+npm dist-tag add @midwayjs/jwt@3.14.7 latest
+npm dist-tag add @midwayjs/mikro@3.14.7 latest
+npm dist-tag add @midwayjs/oss@3.14.9 latest
+npm dist-tag add @midwayjs/passport@3.14.7 latest
+npm dist-tag add @midwayjs/security@3.14.7 latest
+npm dist-tag add @midwayjs/swagger@3.14.9 latest
+npm dist-tag add @midwayjs/tags@3.14.7 latest
+npm dist-tag add @midwayjs/upload@3.14.7 latest
+npm dist-tag add @midwayjs/validate@3.14.7 latest
+npm dist-tag add @midwayjs/version@3.14.9 latest
+npm dist-tag add @midwayjs/express@3.14.7 latest
+
+# Changes:
+
+#  - @midwayjs/axios: 3.14.9 => 3.14.10
+#  - @midwayjs/bull-board: 3.14.9 => 3.14.10
+#  - @midwayjs/code-dye: 3.14.7 => 3.14.10
+#  - @midwayjs/consul: 3.14.7 => 3.14.10
+#  - @midwayjs/cos: 3.14.9 => 3.14.10
+#  - @midwayjs/cross-domain: 3.14.7 => 3.14.10
+#  - @midwayjs/express-session: 3.14.7 => 3.14.10
+#  - @midwayjs/http-proxy: 3.14.9 => 3.14.10
+#  - @midwayjs/i18n: 3.14.7 => 3.14.10
+#  - @midwayjs/info: 3.14.7 => 3.14.10
+#  - @midwayjs/jwt: 3.14.7 => 3.14.10
+#  - @midwayjs/mikro: 3.14.7 => 3.14.10
+#  - @midwayjs/oss: 3.14.9 => 3.14.10
+#  - @midwayjs/passport: 3.14.7 => 3.14.10
+#  - @midwayjs/security: 3.14.7 => 3.14.10
+#  - @midwayjs/swagger: 3.14.9 => 3.14.10
+#  - @midwayjs/tags: 3.14.7 => 3.14.10
+#  - @midwayjs/upload: 3.14.7 => 3.14.10
+#  - @midwayjs/validate: 3.14.7 => 3.14.10
+#  - @midwayjs/version: 3.14.9 => 3.14.10
+#  - @midwayjs/express: 3.14.7 => 3.14.10